RDKB-7592: Extender device metric report is shows devicemac as interface mac
Reason for change: The 'interface_mac' was set to 'device_mac' for extender
connected devices in InterfaceDevicesWifiExtender report.
The expectation is that the 'interface_mac' should be set to bssid of
extender network interface to which the device is connected to.

Test Procedure: Boot the CPE. Connect an extender to it.
Then connect some devices to the extender.Enable InterfaceWifiExtender report.
Check the interface_mac field for each connected device in the report.

Change-Id: I920d5d16f3503dd10109250897807d87c1d640e1
Risks: None
Signed-off-by: Roopali_Galgali@cable.comcast.com