RDKB-9258 : All reports are getting accelerated after reboot
Reason for change: After TTL period is reached, and followed by a
CPE reboot, the polling and reporting periods for all reports
were not retained.These values were not persisted in this scenario.

Test Procedure: Enable all reports. Modify the polling and reporting
periods. Wait till TTL period, and verify that the both the periods
are reset to default values. Reboot the CPE. Verify both the periods
again. They should be same as before the reboot.

Change-Id: Ie0f3ae3ec49d385db1fb73f9e4bcb7e810866d58
Risks: None
Signed-off-by: Roopali_Galgali@cable.comcast.com