RDKBDEV-800 CcspLMLite: Device Refresh Event not Sending Correct Status
Reason for change: 1. We are getting multiple events of Online even when device is connected
2. We are getting the status as Connected only when the device is connected for the first time to the CPE,
afterwards we only get status as Active every time we reconnect the device.
To get the status as Connected again, we need to restart the CPE.
As per the fix, since frequent polls are removed,
the notification for reconnected client rely on Online event instead of Connected event.

Test Procedure: Updated in JIRA description
Risks: None

Signed-off-by: Mohamed Shaikh <mohamed.shaikh@t-systems.com>
Change-Id: Ia914a49463a5d26e203375cfc4c683942f08040d