RDKB-39178 - Removed the unused logging variable
Reason for change: Removed the writelog function with a new Ccsplog3 function.
Test Procedure: Build and verified
Risks: Low
Signed-off-by: Gururaaja ESR <gururaja_erodesriranganramlingham@comcast.com>

RDKB-39178,RDKB-40048,SKYH4-5925,RDKB-40358: LMLite writting to WiFiLog.txt.0

Reason for change: Previously the LMLite was posting to LogAgent and but logAgent just drops it; Doesn't get printed to any log file.
With the new change under RDKB-39178, the log was posted to WIFI which is taking over WiFiLog.txt.0 and wifi was given with WiFiLog.txt.1
Now, we ensure that the LMLite log goes to LM.txt.0 (Its still different behaviour compared to 4.12/4.14; where it is not printed antwhere; just dropped)
Test Procedure: Flash the latest build with fix and verify WIFI logs in WiFiLog.txt.0
Risks: Medium
Signed-off-by: Karunakaran A <karunakaran_amirthalingam@cable.comcast.com>

Change-Id: Ib72562fbf0c2add6ef7ce7118211e817285eba8b