update hostname for reserved lan ip addresses ( OFW-4803 F3896SI-2116 )
[Issue]
Device name is displayed as "unknown" in Connected devices page when client gets reserved IP address

[Description]
When a client is connected with reserve ip rules, the ARP is happend when we create a new entry in the hostlist.
By default the hostname value is set as unkown and later in Hosts_SyncDHCP the actual value is updated from the dnsmasq.lease file.
Since the client IP is already updated in ARP while adding host entry, Hosts_SyncDHCP will not triggered in reserve ip case.

[Solution]
Added conditional check to update the Hostname from dnsmasq.lease file if the host name is set as "unknown".

[TESTING]
Steps:

    1. Factory reset the DUT using GUI
    2. Login to GUI-> Advanced settings ->Wireless ->Security
    3. Change 2.4G SSID name and 5G SSID name in UI
    4. In DHCP page, add two reserved rule using wireless client mac address
    5. Connect both private 2.4G and 5G to the respective wireless clients and verify that it gets reserved IP as per rule
    6. Verify device name and IP in connected devices page

Observed Results:

     As per the device information ,Device name and IP are updated in the Connected devices page.

Upstreamable : Yes

Signed-off-by: arudas <arudas.ext@libertyglobal.com>