fix device name is displayed as * in Connected devices page ( OFW-5313 )
[ISSUE DESCRIPTION]
Device name is displayed as * in Connected devices page.

[FIX DESCRIPTION]
If the hostname field is left empty or displayed as an asterisk (*), it generally suggests that the client didn't specify its hostname. While getting the HostName as * there was also a new line ('\n')
which was getting added along with '\0', so made changes to replace '\n' with '\0'.

[TESTING PROCEDURE]
1) Load with latest image in dual stack mode.
2) Navigate to DHCP page in local UI and disable the DHCPv4 server
3) Connect a WiFi client to 5ghz
4) Enable DHCPv4 server from DHCP page.
5) Check the device details in Connected devices Page.
   Device Name should display as Unknown