fix connected devices page in GUI shows Lan private IP with configured vmb subnet ( OFW-5840 F3896SI-2674 )
[Issue]

When multiple LAN clients are connected and  static IP is configured in one Lan Client (VMB Mode), t
he other Lan Clients entry with private IP  is still visible in GUI (Connected devices page).

[Description]

The LAN Client entries with private IP are still present in Connected Devices page and Device.Hosts table still
shows the LAN Private IP as active. For connected devices details we are checking ARP_CACHE_FILE and updating according to this.
During VMB mode we not getting private ip details in ARP_CACHE_FILE and not updating it.
So Device.Hosts table still shows the LAN Private IP as active.

[Solution]

Examine connected clients are present in ARP_CACHE_FILE or not. If not, makes Device.Hosts table shows the LAN Private IP as not active.

[Testing]

1. Upgrade the image with new change.
2. Connect multiple LAN Clients to the modem.
3. Navigate to Connected Devices Page and note all the LAN Client entries present with private IP.
4. Enable the modem with Multi static VMB either from bootfile or ACS.
5. Assign static IP to one of the LAN clients with the same subnet range as brlan0
   and do not configure static IP to other LAN clients.
6. Device.Hosts table should be updated and the Connected Devices page should change accordingly.

Signed-off-by: Anagha prakash <apadincharachalukulangara.ext@libertyglobal.com>